+++ CHANGES.md
@@ -4,7 +4,7 @@ moodle-theme_boost_union
 Changes
 -------
 
-### Unreleased
+### v4.0-r5
 
 * 2022-10-18 - Feature: Filearea for custom fonts, helps to solve #38.
 * 2022-10-15 - Feature: Filearea for additional resources, solves #113.